requirements.txt
tensorflow
nltk
scikit-learn
pandas 
gdown
dvc
mlflow==2.16.0
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure
joblib
types-PyYAML
scipy
Flask
Flask-Cors
transformers
safetensors
torch
requests
pytest
black
keras
-e .